+++ vod-20170321/alibabacloud_vod20170321/models/_upload_media_by_urlrequest.py
@@ -8,6 +8,8 @@ class UploadMediaByURLRequest(DaraModel):
     def __init__(
         self,
         app_id: str = None,
+        enable_first_frame_cover: bool = None,
+        generate_thumbnail: bool = None,
         session_id: str = None,
         storage_location: str = None,
         template_group_id: str = None,
@@ -18,6 +20,8 @@ class UploadMediaByURLRequest(DaraModel):
     ):
         # The ID of the application. Default value: **app-1000000**. For more information, see [Overview](https://help.aliyun.com/document_detail/113600.html).
         self.app_id = app_id
+        self.enable_first_frame_cover = enable_first_frame_cover
+        self.generate_thumbnail = generate_thumbnail
         # The custom identifier for deduplication. If you specify this parameter and send a request, an error is returned if a request with the same identifier was sent in the last 10 minutes. A custom identifier can be up to 50 characters in length and can contain letters, digits, hyphens (-), and underscores (_). If you do not specify this parameter or leave this parameter empty, duplicate requests are not filtered.
         self.session_id = session_id
         # The storage address of the media file.
@@ -74,6 +78,12 @@ class UploadMediaByURLRequest(DaraModel):
         if self.app_id is not None:
             result['AppId'] = self.app_id
 
+        if self.enable_first_frame_cover is not None:
+            result['EnableFirstFrameCover'] = self.enable_first_frame_cover
+
+        if self.generate_thumbnail is not None:
+            result['GenerateThumbnail'] = self.generate_thumbnail
+
         if self.session_id is not None:
             result['SessionId'] = self.session_id
 
@@ -101,6 +111,12 @@ class UploadMediaByURLRequest(DaraModel):
         m = m or dict()
         if m.get('AppId') is not None:
             self.app_id = m.get('AppId')
+
+        if m.get('EnableFirstFrameCover') is not None:
+            self.enable_first_frame_cover = m.get('EnableFirstFrameCover')
+
+        if m.get('GenerateThumbnail') is not None:
+            self.generate_thumbnail = m.get('GenerateThumbnail')
 
         if m.get('SessionId') is not None:
             self.session_id = m.get('SessionId')